Alien Movement & Map
Multiple aliens spawn on a path and move across it. If they get to the base they disappear and a life is subtracted. Towers shoot these aliens and deal dmg. When the health of an alien reaches zero, the alien is removed and the game cash increased. The game map is now defined as its own object. Currently three waves are spawned (5, 10, 15 drones). 

Note, the game updates far quicker since a new bitmap is not being assigned each time the game map is being drawn. A costly procedure.